added more t6 weapons

+++ core/src/main/java/com/zrp200/rkpd2/actors/mobs/Statue.java
@@ -21,6 +21,8 @@
 
 package com.zrp200.rkpd2.actors.mobs;
 
+import com.watabou.utils.Bundle;
+import com.watabou.utils.Random;
 import com.zrp200.rkpd2.Dungeon;
 import com.zrp200.rkpd2.actors.Char;
 import com.zrp200.rkpd2.items.Generator;
@@ -32,8 +34,6 @@ import com.zrp200.rkpd2.journal.Notes;
 import com.zrp200.rkpd2.messages.Messages;
 import com.zrp200.rkpd2.sprites.StatueSprite;
 import com.zrp200.rkpd2.utils.GLog;
-import com.watabou.utils.Bundle;
-import com.watabou.utils.Random;
 
 public class Statue extends Mob {
 	
@@ -46,7 +46,7 @@ public class Statue extends Mob {
 		properties.add(Property.INORGANIC);
 	}
 	
-	protected Weapon weapon;
+	public Weapon weapon;
 	
 	public Statue() {
 		super();
@@ -136,8 +136,10 @@ public class Statue extends Mob {
 	
 	@Override
 	public void die( Object cause ) {
-		weapon.identify();
-		Dungeon.level.drop( weapon, pos ).sprite.drop();
+		if (weapon != null) {
+			weapon.identify();
+			Dungeon.level.drop(weapon, pos).sprite.drop();
+		}
 		super.die( cause );
 	}
